@@ -8,7 +8,7 @@
     <artifactId>newtrig</artifactId>
     <version>1.0-SNAPSHOT</version>
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
         <project.build.encoding>UTF-8</project.build.encoding>
