@@ -1,9 +1,7 @@
 package gg.xp.xivsupport.models;
 
-public interface CurrentMaxPredicted {
-	long getCurrent();
+public interface CurrentMaxPredicted extends CurrentMaxPair {
 
 	long getPredicted();
 
-	long getMax();
 }
